countryinfo/requirements.txt
Django>=5.2,<6
djangorestframework
requests
